Reports Export Fix for CSV Output
The current release of AOR does not enclose strings in double quotes.
This causes parsing issues, especially with Excel, in the form of
misaligned columns. My fix checks to see if the data value to be
exported is numeric. If the value is not numeric, then the value is
enclosed by double-quotes otherwise it is output as is.